+++ apps/ymir-sdl3/src/app/events/gui_event.hpp
@@ -47,6 +47,7 @@ struct PeripheralBindsParams {
 struct GUIEvent {
     enum class Type {
         LoadDisc,
+        LoadRecommendedGameCartridge,
         OpenBackupMemoryCartFileDialog,
         OpenROMCartFileDialog,
         OpenPeripheralBindsEditor,