+++ apps/ymir-sdl3/src/app/profile.hpp
@@ -8,7 +8,7 @@ namespace app {
 enum class ProfilePath {
     Root,            // Root of the profile    <profile>/
     IPLROMImages,    // IPL ROM images         <profile>/roms/ipl/
-    CartROMImages,   // Cartridge ROMs         <profile>/roms/cart/
+    ROMCartImages,   // Cartridge ROMs         <profile>/roms/cart/
     BackupMemory,    // Backup memory images   <profile>/backup/
     ExportedBackups, // Exported backup files  <profile>/backup/exported/
     PersistentState, // Persistent app state   <profile>/state/