+++ apps/ymir-sdl3/src/app/settings.hpp
@@ -267,6 +267,8 @@ public:
         struct ROM {
             std::filesystem::path imagePath;
         } rom;
+
+        bool autoLoadGameCarts;
     } cartridge;
 
     // ---------------------------------------------------------------------------------------------