+++ libs/ymir-core/include/ymir/core/hash.hpp
@@ -26,3 +26,16 @@ XXH128Hash CalcHash128(const void *input, size_t len, uint64_t seed = 0);
 std::string ToString(const XXH128Hash &hash);
 
 } // namespace ymir
+
+template <>
+struct std::hash<ymir::XXH128Hash> {
+    size_t operator()(const ymir::XXH128Hash &e) const noexcept {
+        uint64_t low = 0;
+        uint64_t high = 0;
+        for (size_t i = 0; i < 8; i++) {
+            low |= static_cast<uint64_t>(e[i]) << (i * 8ull);
+            high |= static_cast<uint64_t>(e[i + 8]) << (i * 8ull);
+        }
+        return std::hash<uint64_t>{}(low) ^ std::hash<uint64_t>{}(high);
+    }
+};